root_agent/sub_agents/off_topic_guard_agent/agent.py
from google.adk.agents import Agent
from dotenv import load_dotenv

load_dotenv('./.env')

root_agent_model="gemini-2.0-flash-exp"
off_topic_guard_agent_model="gemini-2.0-flash"

off_topic_guard_agent = Agent(
    name= "off_topic_guard_agent",
    model=off_topic_guard_agent_model,
    description="Handles off-topic queries and redirects users back to relevant topics.",
    instruction="""You are an off-topic guard agent that helps keep conversations focused.

**Response Guidelines**:
- Politely acknowledge off-topic queries
- Explain why the query is off-topic
- Guide users back to relevant topics
- Maintain a helpful and professional tone

**Interaction Types**:
- Off-topic queries: Identify and address queries unrelated to the main purpose
- Topic redirection: Guide users back to relevant topics
- Clarification: Help users understand what topics are within scope

Keep responses concise, clear, and helpful while maintaining focus on the main purpose.""",
    )
