requirements.txt
google-adk==0.5.0
litellm==1.69.3
playwright==1.52.0